+++ ecs/entity.go
@@ -6,10 +6,10 @@ import "reflect"
 var entityType = reflect.TypeOf(Entity{})
 
 // Size of an [Entity] in memory, in bytes.
-var entitySize = entityType.Size()
+var entitySize uint32 = uint32(entityType.Size())
 
 // Size of an [entityIndex] in memory.
-var entityIndexSize = reflect.TypeOf(entityIndex{}).Size()
+var entityIndexSize uint32 = uint32(reflect.TypeOf(entityIndex{}).Size())
 
 // Entity identifier.
 // Holds an entity ID and it's generation for recycling.
@@ -41,5 +41,5 @@ func (e Entity) IsZero() bool {
 // entityIndex indicates where an entity is currently stored.
 type entityIndex struct {
 	arch  *archetype // Entity's current archetype
-	index uintptr    // Entity's current index in the archetype
+	index uint32     // Entity's current index in the archetype
 }